Fixes and Changes:
Fixed Secondary and Target animations failing when using the Side Impact option

Changed Hooks.callAll("AutomatedAnimations-WorkflowStart", data); to also allow passing recheckAnimation. This is useful if you wish to change the item being processed before continuing the workflow. If this is used, it will fall back to the original item animation data if none is matched.